Removed extraneous root template
